Add new Logo
Copyright free picture sourced from Pixabay

https://pixabay.com/illustrations/google-maps-navigation-gps-maps-1797882/

## Pixabay License
License for Content – Pixabay License
Content on Pixabay is made available to you on the following terms ("Pixabay License"). Under the Pixabay License you are granted an irrevocable, worldwide, non-exclusive and royalty free right to use, download, copy, modify or adapt the Content for commercial or non-commercial purposes. Attribution of the photographer, videographer, musician or Pixabay is not required but is always appreciated.

The Pixabay License does not allow:

Sale or distribution of Content as digital Content or as digital wallpapers (such as on stock media websites);
Sale or distribution of Content e.g. as a posters, digital prints, music files or physical products, without adding any additional elements or otherwise adding value
Depiction of identifiable persons in an offensive, pornographic, obscene, immoral, defamatory or libelous way; or
Any suggestion that there is an endorsement of products and services by depicted persons, brands, vocalists and organisations, unless permission was granted.
Please be aware that while all Content on Pixabay is free to use for commercial and non-commercial purposes, items in the Content, such as identifiable people, logos, brands, audio samples etc. may be subject to additional copyrights, property rights, privacy rights, trademarks etc. and may require the consent of a third party or the license of these rights - particularly for commercial applications. Pixabay does not represent or warrant that such consents or licenses have been obtained, and expressly disclaims any liability in this respect.